✅  Move some tests to the Input component
